requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose
cloudinary
urllib3
pymysql


